+++ client/driver/logging/collector_windows.go
@@ -1,4 +1,4 @@
-package logcollector
+package logging
 
 import (
 	"log"
