+++ client/driver/logging/syslog_server.go
@@ -1,4 +1,4 @@
-package logcollector
+package logging
 
 import (
 	"bufio"
