+++ client/driver/logging/universal_collector.go
@@ -1,6 +1,6 @@
 // +build !windows
 
-package logcollector
+package logging
 
 import (
 	"fmt"
